move all declarations into separate files
