+++ godtools/App/Features/Tools/Domain/GetBannerImageUseCase/DefaultGetBannerImageUseCase.swift
@@ -1,19 +1,12 @@
 //
-//  GetBannerImageUseCase.swift
+//  DefaultGetBannerImageUseCase.swift
 //  godtools
 //
-//  Created by Rachael Skeath on 4/11/22.
+//  Created by Rachael Skeath on 4/25/22.
 //  Copyright © 2022 Cru. All rights reserved.
 //
 
-import Foundation
 import SwiftUI
-
-protocol GetBannerImageUseCase {
-    func getBannerImage() -> Image?
-}
-
-// MARK: - Default
 
 class DefaultGetBannerImageUseCase: GetBannerImageUseCase {
     
@@ -39,22 +32,5 @@ class DefaultGetBannerImageUseCase: GetBannerImageUseCase {
         else {
             return nil
         }
-    }
-}
-
-// MARK: - Mock
-
-class MockGetBannerImageUseCase: GetBannerImageUseCase {
-    
-    var nilImage: Bool = false
-    
-    private static let bannerImageNames = ["banner1", "banner2"]
-    
-    init(nilImage: Bool = false) {
-        self.nilImage = nilImage
-    }
-    
-    func getBannerImage() -> Image? {
-        return nilImage ? nil : Image(MockGetBannerImageUseCase.bannerImageNames.randomElement()!)
     }
 }
