+++ godtools/App/Features/Tools/Domain/GetBannerImageUseCase/GetBannerImageUseCase.swift
@@ -1,0 +1,13 @@
+//
+//  GetBannerImageUseCase.swift
+//  godtools
+//
+//  Created by Rachael Skeath on 4/11/22.
+//  Copyright © 2022 Cru. All rights reserved.
+//
+
+import SwiftUI
+
+protocol GetBannerImageUseCase {
+    func getBannerImage() -> Image?
+}
