+++ godtools/App/Features/Tools/Domain/GetToolDataUseCase/DefaultGetToolDataUseCase.swift
@@ -1,24 +1,12 @@
 //
-//  GetToolDataUseCase.swift
+//  DefaultGetToolDataUseCase.swift
 //  godtools
 //
-//  Created by Rachael Skeath on 4/11/22.
+//  Created by Rachael Skeath on 4/25/22.
 //  Copyright © 2022 Cru. All rights reserved.
 //
 
 import Foundation
-
-protocol GetToolDataUseCase {
-    func getToolData() -> ToolDataModel
-}
-
-struct ToolDataModel {
-    let title: String
-    let category: String
-    let languageDirection: LanguageDirection
-}
-
-// MARK: - Default
 
 class DefaultGetToolDataUseCase: GetToolDataUseCase {
     
@@ -65,19 +53,5 @@ class DefaultGetToolDataUseCase: GetToolDataUseCase {
             category: localizationServices.stringForBundle(bundle: languageBundle, key: "tool_category_\(resource.attrCategory)"),
             languageDirection: languageDirection
         )
-    }
-}
-
-// MARK: - Mock
-
-class MockGetToolDataUseCase: GetToolDataUseCase {
-    let languageDirection: LanguageDirection
-    
-    init(languageDirection: LanguageDirection) {
-        self.languageDirection = languageDirection
-    }
-    
-    func getToolData() -> ToolDataModel {
-        return ToolDataModel(title: "Tool Data Title", category: "Tool Category", languageDirection: languageDirection)
     }
 }
