+++ godtools/App/Features/Tools/Domain/GetToolDataUseCase/GetToolDataUseCase.swift
@@ -1,0 +1,13 @@
+//
+//  GetToolDataUseCase.swift
+//  godtools
+//
+//  Created by Rachael Skeath on 4/11/22.
+//  Copyright © 2022 Cru. All rights reserved.
+//
+
+import Foundation
+
+protocol GetToolDataUseCase {
+    func getToolData() -> ToolDataModel
+}
