+++ godtools/App/Features/Tools/Domain/GetToolDataUseCase/ToolDataModel.swift
@@ -1,0 +1,15 @@
+//
+//  ToolDataModel.swift
+//  godtools
+//
+//  Created by Rachael Skeath on 4/25/22.
+//  Copyright © 2022 Cru. All rights reserved.
+//
+
+import Foundation
+
+struct ToolDataModel {
+    let title: String
+    let category: String
+    let languageDirection: LanguageDirection
+}
