+++ godtools/App/Flows/MockFlowDelegate.swift
@@ -1,0 +1,13 @@
+//
+//  MockFlowDelegate.swift
+//  godtools
+//
+//  Created by Rachael Skeath on 4/25/22.
+//  Copyright © 2022 Cru. All rights reserved.
+//
+
+import Foundation
+
+class MockFlowDelegate: FlowDelegate {
+    func navigate(step: FlowStep) { }
+}
